Added 'meetsGoalConditions' to FluxxCard (incomplete).
meetsGoalConditions takes a goal card and player. It runs through all the goal cards, comparing to see if the player has met the goal or not. It will not check for Creepers. That will take place in isWinner(), because FluxxGame can check the rules to see if cards like Silver Lining or Needs Potato are there.